feat: use server api
