+++ layouts/default.vue
@@ -8,7 +8,7 @@ const host = process.server
 
 useHead({
   link: [
-    // We use $route.path since we don't use query parameters
+    // We use route.path since we don't use query parameters
     { rel: 'canonical', href: `https://${host}${route.path}` }
   ]
 })
