+++ server/api/hn/feeds.ts
@@ -1,20 +1,20 @@
-import { createError, PHandle } from 'h3'
-import { $fetch } from 'ohmyfetch/node'
-import { getQuery, parseURL, withoutLeadingSlash } from 'ufo'
+import { createError } from 'h3'
+import { $fetch } from 'ohmyfetch'
+import { getQuery, parseURL } from 'ufo'
 
-import { feeds, validFeeds } from '../../../common/api'
+import { feedsInfo, validFeeds } from '../../../composables/api'
 
-import { baseURL } from '.'
+import { baseURL } from '../../constants'
 
-const feedUrls: Record<keyof typeof feeds, string> = {
+const feedUrls: Record<keyof typeof feedsInfo, string> = {
   ask: 'askstories',
   jobs: 'jobstories',
   show: 'showstories',
   newest: 'newstories',
-  news: 'topstories',
+  news: 'topstories'
 }
 
-async function fetchFeed(feed: string, page = '1') {
+async function fetchFeed (feed: string, page = '1') {
   const { fetchItem } = await import('./item')
   const entries = Object.values(
     await $fetch(`${baseURL}/${feedUrls[feed]}.json`)
@@ -22,21 +22,16 @@ async function fetchFeed(feed: string, page = '1') {
   return Promise.all(entries.map(id => fetchItem(id)))
 }
 
-const handler: PHandle = async req => {
-  const { pathname, search } = parseURL(req.url)
-  const feed = withoutLeadingSlash(pathname)
-  const { page = '1' } = getQuery(search)
+export default defineEventHandler(({ req }) => {
+  const { search } = parseURL(req.url)
+  const { page = '1', feed = 'news' } = getQuery(search) as { page: string, feed: string }
 
   if (!validFeeds.includes(feed) || String(Number(page)) !== page) {
     throw createError({
       statusCode: 422,
-      statusMessage: `Must provide one of ${validFeeds.join(
-        ', '
-      )} and a valid page number.`,
+      statusMessage: `Must provide one of ${validFeeds.join(', ')} and a valid page number.`
     })
   }
 
   return fetchFeed(feed, page)
-}
-
-export default handler
+})
