+++ server/api/hn/item.ts
@@ -1,23 +1,10 @@
-import { createError, PHandle } from 'h3'
-import { $fetch } from 'ohmyfetch/node'
+import { createError } from 'h3'
+import { $fetch } from 'ohmyfetch'
 import { withoutLeadingSlash } from 'ufo'
+import { baseURL } from '../../constants'
+import { Item } from '~/types'
 
-import { baseURL } from '.'
-
-export interface Item {
-  id: number
-  url?: string
-  title?: string
-  type: 'job' | 'story' | 'comment' | 'poll'
-  points: number
-  user: string
-  content?: string
-  time: string
-  comments_count?: number
-  comments?: Item[]
-}
-
-export async function fetchItem(
+export async function fetchItem (
   id: string,
   withComments = false
 ): Promise<Item> {
@@ -35,23 +22,21 @@ export async function fetchItem(
     comments_count: Object.values(item.kids).length,
     comments: withComments
       ? await Promise.all(
-          Object.values(item.kids as string[]).map(id =>
-            fetchItem(id, withComments)
-          )
+        Object.values(item.kids as string[]).map(id =>
+          fetchItem(id, withComments)
         )
-      : [],
+      )
+      : []
   }
 }
 
-const handler: PHandle = async req => {
+export default defineEventHandler(({ req }) => {
   const itemId = withoutLeadingSlash(req.url)
   if (!itemId) {
     throw createError({
       statusCode: 422,
-      statusMessage: 'Must provide a user ID.',
+      statusMessage: 'Must provide a user ID.'
     })
   }
   return fetchItem(itemId, true)
-}
-
-export default handler
+})
