--- server/api/swr.ts
@@ -1,4 +1,0 @@
-export default (_, res, next) => {
-  res.setHeader('Cache-Control', 's-maxage=10, stale-while-revalidate')
-  next()
-}
