+++ server/constants.ts
@@ -1,0 +1,1 @@
+export const baseURL = 'https://hacker-news.firebaseio.com/v0'
