+++ types.ts
@@ -1,0 +1,19 @@
+export interface Item {
+  id: number
+  url?: string
+  title?: string
+  type: 'job' | 'story' | 'comment' | 'poll'
+  points: number
+  user: string
+  content?: string
+  time: string
+  comments_count?: number
+  comments?: Item[]
+}
+
+export interface User {
+  id: string
+  created_time: string
+  karma: number
+  about: string
+}
